moves profile model from project app to user_profile app and renames that model to User_profile and also updates apps in setting and adds cors in middleware.
